feat: update dependencies and restructure application layout
- Updated `next` to version 15.5.2 and added `next-intl` for internationalization support.
- Removed outdated pages and components from the `src/app` directory, including affiliates, clash-shop, contact, faq, and rules pages.
- Introduced new layout structure for main application pages, ensuring consistent header and footer across all pages.
- Added internationalization support with message files for English and Bengali.
- Implemented new `ContactSection` and `FaqHeroSection` components with updated styles and structure.
- Refactored `HeroSection` in rules to utilize translations for dynamic content.
