+++ messages/bn.json
@@ -1,0 +1,5 @@
+{
+  "RulesPage": {
+    "title": "বাংলা নিয়মাবলী"
+  }
+}
