+++ messages/en.json
@@ -1,0 +1,5 @@
+{
+  "RulesPage": {
+    "title": "Trading rules"
+  }
+}
