+++ src/app/(main)/affiliates/page.tsx
@@ -1,9 +1,9 @@
 import WhyChooseProgram from "@/components/affiliates/WhyChoosePrograms";
-import HeroSection from "../../components/affiliates/HeroSection";
+import HeroSection from "../../../components/affiliates/HeroSection";
 
-import HowItWorksAffiliates from "../../components/affiliates/HowItWorksAffiliates";
+import HowItWorksAffiliates from "../../../components/affiliates/HowItWorksAffiliates";
 
-import Banner from "../../components/Banner";
+import Banner from "../../../components/Banner";
 
 import TiredRewards from "@/components/affiliates/TiredRewards";
 import CommissionSection from "@/components/affiliates/CommissionSection";
