+++ src/app/(main)/clash-shop/page.tsx
@@ -1,13 +1,13 @@
 // app/rules/page.tsx (Next.js 13+ App Router)
 
-import FeaturesSection from "../../components/FeaturesSection";
-import HeroSection from "../../components/clash-shop/HeroSection";
-import SpinDetailsSection from "../../components/clash-shop/SpinDetailsSection";
-import HowItWorks from "../../components/HowItWorks";
-import Subscriptions from "../../components/Subscriptions";
-import SpinWinTrade from "../../components/SpinWinTrade";
-import Faq from "../../components/Faq";
-import Banner from "../../components/Banner";
+import FeaturesSection from "../../../components/FeaturesSection";
+import HeroSection from "../../../components/clash-shop/HeroSection";
+import SpinDetailsSection from "../../../components/clash-shop/SpinDetailsSection";
+import HowItWorks from "../../../components/HowItWorks";
+import Subscriptions from "../../../components/Subscriptions";
+import SpinWinTrade from "../../../components/SpinWinTrade";
+import Faq from "../../../components/Faq";
+import Banner from "../../../components/Banner";
 import RulesSection from "@/components/rules/RulesSection";
 import HowToEarnCoins from "@/components/clash-shop/HowToEarnCoins";
