+++ src/app/(main)/faq/FaqHeroSection.tsx
