+++ src/app/(main)/page.tsx
@@ -1,13 +1,13 @@
 // app/page.js (Next.js 13+ App Router)
 // or pages/index.js (Pages Router)
 
-import FeaturesSection from "../components/FeaturesSection";
-import HeroSection from "../components/HeroSection";
-import HowItWorks from "../components/HowItWorks";
-import Subscriptions from "../components/Subscriptions";
-import SpinWinTrade from "../components/SpinWinTrade";
-import Faq from "../components/Faq";
-import Banner from "../components/Banner";
+import FeaturesSection from "../../components/FeaturesSection";
+import HeroSection from "../../components/HeroSection";
+import HowItWorks from "../../components/HowItWorks";
+import Subscriptions from "../../components/Subscriptions";
+import SpinWinTrade from "../../components/SpinWinTrade";
+import Faq from "../../components/Faq";
+import Banner from "../../components/Banner";
 
 export default function Home() {
   return (
