+++ src/app/(main)/rules/page.tsx
@@ -1,12 +1,12 @@
 // app/rules/page.tsx (Next.js 13+ App Router)
 
-import FeaturesSection from "../../components/FeaturesSection";
-import HeroSection from "../../components/rules/HeroSection";
-import HowItWorks from "../../components/HowItWorks";
-import Subscriptions from "../../components/Subscriptions";
-import SpinWinTrade from "../../components/SpinWinTrade";
-import Faq from "../../components/Faq";
-import Banner from "../../components/Banner";
+import FeaturesSection from "../../../components/FeaturesSection";
+import HeroSection from "../../../components/rules/HeroSection";
+import HowItWorks from "../../../components/HowItWorks";
+import Subscriptions from "../../../components/Subscriptions";
+import SpinWinTrade from "../../../components/SpinWinTrade";
+import Faq from "../../../components/Faq";
+import Banner from "../../../components/Banner";
 import RulesSection from "@/components/rules/RulesSection";
 
 export default function RulesPage() {
