+++ src/app/login/layout.tsx
@@ -1,21 +1,10 @@
 import type { Metadata } from "next";
-import { Geist, Geist_Mono } from "next/font/google";
-import "../globals.css";
+
 import LoginNavbar from "../LoginNavbar";
 
-const geistSans = Geist({
-  variable: "--font-geist-sans",
-  subsets: ["latin"],
-});
-
-const geistMono = Geist_Mono({
-  variable: "--font-geist-mono",
-  subsets: ["latin"],
-});
-
 export const metadata: Metadata = {
-  title: "Create Next App",
-  description: "Generated by create next app",
+  title: "Login - SoloClash",
+  description: "Login to your SoloClash account",
 };
 
 export default function RootLayout({
@@ -24,14 +13,16 @@ export default function RootLayout({
   children: React.ReactNode;
 }>) {
   return (
-    <>
-      {" "}
-      {/* Navigation Header */}
-      <header className="absolute top-10 w-full z-50">
-        <LoginNavbar />
-      </header>
-      {/* Main content */}
-      {children}
-    </>
+    <html lang="en">
+      <body>
+        {" "}
+        {/* Navigation Header */}
+        <header className="absolute top-10 w-full z-50">
+          <LoginNavbar />
+        </header>
+        {/* Main content */}
+        {children}
+      </body>
+    </html>
   );
 }
