+++ src/app/signup/layout.tsx
@@ -1,22 +1,10 @@
 import type { Metadata } from "next";
-import { Geist, Geist_Mono } from "next/font/google";
-import "../globals.css";
-import LoginNavbar from "../LoginNavbar";
+
 import SignupNavbar from "../SignupNavbar";
 
-const geistSans = Geist({
-  variable: "--font-geist-sans",
-  subsets: ["latin"],
-});
-
-const geistMono = Geist_Mono({
-  variable: "--font-geist-mono",
-  subsets: ["latin"],
-});
-
 export const metadata: Metadata = {
-  title: "Create Next App",
-  description: "Generated by create next app",
+  title: "Signup - SoloClash",
+  description: "Create your SoloClash account",
 };
 
 export default function RootLayout({
@@ -26,12 +14,10 @@ export default function RootLayout({
 }>) {
   return (
     <>
-      {" "}
-      {/* Navigation Header */}
       <header className="absolute top-10 w-full z-50">
         <SignupNavbar />
       </header>
-      {/* Main content */}
+
       {children}
     </>
   );
