+++ src/i18n/request.ts
@@ -1,0 +1,11 @@
+import { getRequestConfig } from "next-intl/server";
+
+export default getRequestConfig(async () => {
+  // Static for now, we'll change this later
+  const locale = "bn";
+
+  return {
+    locale,
+    messages: (await import(`../../messages/${locale}.json`)).default,
+  };
+});
